+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/models/ModuleDescriptor.kt
@@ -30,3 +30,6 @@ data class ModuleDescriptor(
         )
     }
 }
+
+internal inline val ModuleDescriptor.path
+    get() = resources?.resourceRoot?.path?.removePrefix("../")